requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2024.1
diff-match-patch==20230430
distlib==0.3.7
Django==4.2.9
django-admin-adminlte1.0.5
django-ckeditor==6.7.1
django-crispy-forms2.1
django-extensions==3.0.2
django-filter==24.2
django-import-export4.0.7
django-js-asset==2.2.0
django-social-share2.3.0
django-widget-tweaks1.5.0
filelock==3.12.2
graphviz==0.14
gunicorn==21.2.0
mysql-connector-python8.0.20
mysqlclient==2.2.0
packaging==24.0
pillow==10.3.0
pip==24.0
platformdirs==3.10.0
protobuf==3.12.2
pydot==1.4.1
PyMySQL==1.1.0
pyparsing==2.4.7
python-dotenv==1.0.1
pytz==2020.1
setuptools==65.5.0
six==1.15.0
sqlparse==0.3.1
str2bool==1.1
tablib==3.5.0
tzdata==2023.3
virtualenv==20.24.2
virtualenvwrapper-win==1.2.7
whitenoise==6.6.0